Extract aidev_lib/routing.py (#17 Stage 3)
Move classification and model-selection helpers out of aidev.py into
a new aidev_lib/routing.py: classify_task, choose_model,
normalize_model_name, supervisor_policy, plus the helpers used only
by them (model_catalog_from, normalize_tags). aidev.py now re-exports
those names so `from aidev import classify_task, ...` keeps working
unchanged for tests and external callers.

Naming note: the new package is `aidev_lib/` (not `aidev/`) on
purpose. Putting `aidev/` next to `aidev.py` would shadow the
module — Python's import system prefers a same-named package over
a module — and would break every `from aidev import ...` site.
`aidev_lib/` preserves the script invocation, the wrapper scripts
(aidev.ps1, AI Dev.cmd), and the existing test suite without
touching them. A future stage may reconsider this once enough
material has moved that an `aidev/__main__.py` flip is worth the
UX disruption.

Pure refactor: behavior is byte-equivalent. The fallback
`models` dict that `choose_model`/`supervisor_policy` use when
the caller's config is missing is duplicated as a small private
constant `_DEFAULT_MODELS` inside routing.py to avoid a circular
import on `aidev.DEFAULT_CONFIG` at script-vs-package load time.
Values match `DEFAULT_CONFIG["models"]` exactly.

aidev.py shrinks from 1564 to 1349 lines. aidev_lib/routing.py
is 245 lines.

Verified:
- python -m pytest tests/  ->  71 passed
- python aidev.py "Access blocked Authorization Error OAuth client
  was not found Error 401 invalid_client Rewrite" --ui-settings "{}"
  -> Reasoning: medium (regression-tested for #7)
- python aidev.py "hello can you code hello world in python"
  --ui-settings "{}"
  -> Task type: small_create, Reasoning: none
- cd desktop && npm run smoke  ->  Smoke checks passed.

Refs #17
